订单修改金额的类型。返回订单clentOrder 添加金额元的属性。

+++ store-system-common/src/main/java/com/store/system/model/Order.java
@@ -68,10 +68,10 @@ public class Order implements Serializable {
 
     private String desc; //描述
 
-    private double totalPrice;//总金额(分)
+    private int totalPrice;//总金额(分)
     private double discount;//折扣
-    private double dicountPrice;//折后金额(分)
-    private double price;//实际支付金额(分)
+    private int dicountPrice;//折后金额(分)
+    private int price;//实际支付金额(分)
     private long couponid;//营销券id
     private int marketingType;//促销类型
     private int makeStatus;//加工状态/取货状态
